Revert "move receipts types to another file"
This reverts commit 1e783117c36e4fb231ccaaa4ef6b18465ae979af.
